add login hook
